Refatorada classes de veículos, adicionando um arquivo para cada veículo e sua lógica de negócio.
