+++ veiculos/import_relatives.rb
@@ -1,0 +1,6 @@
+#Lista de Veiculos
+
+require_relative 'bicicleta'
+require_relative 'cavalo'
+require_relative 'drone'
+require_relative 'moto'
